vision/requirements.txt
click==8.1.3
dlib==19.24.0
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.24.2
opencv-python==4.7.0.72
Pillow==9.4.0
click==8.1.3
dlib==19.24.0
face-recognition==1.3.0
face-recognition-models==0.3.0
joblib==1.2.0
numpy==1.24.2
opencv-python==4.7.0.72
Pillow==9.4.0
scikit-learn==1.2.2
scipy==1.10.1
threadpoolctl==3.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dlib==19.24.0
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.12.0
fsspec==2023.5.0
huggingface-hub==0.14.1
idna==3.4
joblib==1.2.0
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.1
Pillow==9.4.0
PyYAML==6.0
requests==2.30.0
scikit-learn==1.2.2
scipy==1.10.1
skops==0.6.0
tabulate==0.9.0
threadpoolctl==3.1.0
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==2.0.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dlib==19.24.0
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.12.0
fsspec==2023.5.0
huggingface-hub==0.14.1
idna==3.4
joblib==1.2.0
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.1
Pillow==9.4.0
PyYAML==6.0
requests==2.30.0
scikit-learn==1.2.2
scipy==1.10.1
skops==0.6.0
tabulate==0.9.0
threadpoolctl==3.1.0
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==2.0.2
